service -> server per smithy naming recs
