+++ codegen/smithy-go-codegen/src/main/java/software/amazon/smithy/go/codegen/server/NotImplementedError.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.smithy.go.codegen.service;
+package software.amazon.smithy.go.codegen.server;
 
 import static software.amazon.smithy.go.codegen.GoWriter.goTemplate;
