+++ codegen/smithy-go-codegen/src/main/java/software/amazon/smithy/go/codegen/server/protocol/JsonDeserializerGenerator.java
@@ -13,12 +13,12 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.smithy.go.codegen.service.protocol;
+package software.amazon.smithy.go.codegen.server.protocol;
 
 import static software.amazon.smithy.go.codegen.GoWriter.goTemplate;
 import static software.amazon.smithy.go.codegen.SymbolUtils.getReference;
 import static software.amazon.smithy.go.codegen.SymbolUtils.isPointable;
-import static software.amazon.smithy.go.codegen.service.ServiceCodegenUtils.normalize;
+import static software.amazon.smithy.go.codegen.server.ServerCodegenUtil.normalize;
 
 import java.util.Set;
 import software.amazon.smithy.codegen.core.CodegenException;
