+++ codegen/smithy-go-codegen/src/main/resources/META-INF/services/software.amazon.smithy.build.SmithyBuildPlugin
@@ -1,2 +1,2 @@
 software.amazon.smithy.go.codegen.GoCodegenPlugin
-software.amazon.smithy.go.codegen.service.ServiceCodegenPlugin
+software.amazon.smithy.go.codegen.server.ServerCodegenPlugin
